配置过程不删除启动器

+++ base-remove.sh
@@ -57,12 +57,12 @@ rm -rf /etc/systemd/system/systemd-nspawn@$1.service.d
 rm -f /etc/systemd/nspawn/$1.nspawn
 [[ -d /etc/systemd/nspawn && `ls -A /etc/systemd/nspawn |wc -w` == 0 ]] && rm -rf /home/share /etc/systemd/nspawn
 
-if [ -f /usr/share/applications/deepin-qq.desktop ] && [[ $(cat /usr/share/applications/deepin-qq.desktop | grep $1-) ]]; then
+if [[ ! $EXEC_FROM_CONFIG ]] && [ -f /usr/share/applications/deepin-qq.desktop ] && [[ $(cat /usr/share/applications/deepin-qq.desktop | grep $1-) ]]; then
     rm -f /usr/share/pixmaps/com.qq.im.deepin.svg
     rm -f /usr/share/applications/deepin-qq.desktop
 fi
 
-if [ -f /usr/share/applications/deepin-weixin.desktop ] && [[ $(cat /usr/share/applications/deepin-weixin.desktop | grep $1-) ]]; then
+if [[ ! $EXEC_FROM_CONFIG ]] && [ -f /usr/share/applications/deepin-weixin.desktop ] && [[ $(cat /usr/share/applications/deepin-weixin.desktop | grep $1-) ]]; then
     rm -f /usr/share/pixmaps/com.qq.weixin.deepin.svg
     rm -f /usr/share/applications/deepin-weixin.desktop
 fi
